third_party/WebKit/LayoutTests/inspector/console/console-nested-group-expected.txt

CONSOLE MESSAGE: line 11: Message inside inner group
CONSOLE MESSAGE: line 14: Message that must not be in any group
CONSOLE MESSAGE: line 21: Message inside third group
Tests that console.group/groupEnd messages won't be coalesced. Bug 56114. Bug 63521.

console-nested-group.html:9 outer group console-message console-log-level console-group-title
console-nested-group.html:10 inner group console-message console-log-level console-group-title
console-nested-group.html:11 Message inside inner group console-message console-log-level
console-nested-group.html:14 Message that must not be in any group console-message console-log-level
console-nested-group.html:19 One of several groups which shouldn't be coalesced. console-message console-log-level console-group-title
console-nested-group.html:19 One of several groups which shouldn't be coalesced. console-message console-log-level console-group-title
console-nested-group.html:19 One of several groups which shouldn't be coalesced. console-message console-log-level console-group-title
console-nested-group.html:21 Message inside third group console-message console-log-level
